requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
aiosignal==1.2.0
    # via ray
attrs==21.4.0
    # via
    #   jsonschema
    #   pytest
    #   ray
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   flask
    #   ray
cycler==0.11.0
    # via matplotlib
distlib==0.3.4
    # via virtualenv
filelock==3.7.0
    # via
    #   ray
    #   virtualenv
flask==2.1.2
    # via
    #   -r requirements/requirements.in
    #   flask-classful
flask-classful==0.14.2
    # via -r requirements/requirements.in
fonttools==4.33.3
    # via matplotlib
frozenlist==1.3.0
    # via
    #   aiosignal
    #   ray
gputil==1.4.0
    # via -r requirements/requirements.in
grpcio==1.43.0
    # via ray
idna==3.3
    # via requests
importlib-metadata==4.11.4
    # via flask
iniconfig==1.1.1
    # via pytest
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jsonpickle==2.2.0
    # via -r requirements/requirements.in
jsonschema==4.5.1
    # via ray
kiwisolver==1.4.2
    # via matplotlib
loguru==0.6.0
    # via -r requirements/requirements.in
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.2
    # via -r requirements/requirements.in
msgpack==1.0.3
    # via ray
numpy==1.22.4
    # via
    #   matplotlib
    #   ray
p2pnetwork==1.2
    # via -r requirements/requirements.in
packaging==21.3
    # via
    #   matplotlib
    #   pytest
pillow==9.1.1
    # via matplotlib
platformdirs==2.5.2
    # via virtualenv
pluggy==1.0.0
    # via pytest
protobuf==3.20.1
    # via ray
py==1.11.0
    # via pytest
pycryptodome==3.16.0
    # via -r requirements/requirements.in
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.2
    # via
    #   -r requirements/requirements.in
    #   pytest-asyncio
pytest-asyncio==0.18.3
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via matplotlib
python-dotenv==0.20.0
    # via -r requirements/requirements.in
pyyaml==6.0
    # via ray
# ray==1.12.1
    # via -r requirements/requirements.in
requests==2.27.1
    # via
    #   -r requirements/requirements.in
    #   ray
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
    #   virtualenv
tomli==2.0.1
    # via pytest
urllib3==1.26.9
    # via requests
virtualenv==20.14.1
    # via ray
waitress==2.1.2
    # via -r requirements/requirements.in
werkzeug==2.1.2
    # via flask
zipp==3.8.0
    # via importlib-metadata
pylint==2.15.3
Whoosh==2.7.4
speedtest-cli==2.1.3
psutil==5.9.4
mnemonic
cryptography
environ